+++ Game/src/Game/Scripts/cameraMove.cpp
@@ -37,18 +37,18 @@ void CameraMove::Move()
     //     input.y = -5 * GetFrameTime();
     // }
 
-    if (IsKeyDown(KEY_RIGHT) || IsGamepadButtonDown(0, GAMEPAD_BUTTON_LEFT_FACE_RIGHT)) {
-        input.x = 5 * GetFrameTime();
+    if (IsKeyDown(KEY_D) || IsGamepadButtonDown(0, GAMEPAD_BUTTON_LEFT_FACE_RIGHT)) {
+        input.x = 7.5 * GetFrameTime();
     }
-    if (IsKeyDown(KEY_LEFT) || IsGamepadButtonDown(0, GAMEPAD_BUTTON_LEFT_FACE_LEFT)) {
-        input.x = -5 * GetFrameTime();
+    if (IsKeyDown(KEY_A) || IsGamepadButtonDown(0, GAMEPAD_BUTTON_LEFT_FACE_LEFT)) {
+        input.x = -7.5 * GetFrameTime();
     }
 
-    if (IsKeyDown(KEY_UP) || IsGamepadButtonDown(0, GAMEPAD_BUTTON_LEFT_FACE_UP)) {
-        input.y = 5 * GetFrameTime();
+    if (IsKeyDown(KEY_W) || IsGamepadButtonDown(0, GAMEPAD_BUTTON_LEFT_FACE_UP)) {
+        input.y = 7.5 * GetFrameTime();
     }
-    if (IsKeyDown(KEY_DOWN) || IsGamepadButtonDown(0, GAMEPAD_BUTTON_LEFT_FACE_DOWN)) {
-        input.y = -5 * GetFrameTime();
+    if (IsKeyDown(KEY_S) || IsGamepadButtonDown(0, GAMEPAD_BUTTON_LEFT_FACE_DOWN)) {
+        input.y = -7.5 * GetFrameTime();
     }
 
     velocity = { input.x * moveSpeed, input.y * moveSpeed };
@@ -56,21 +56,22 @@ void CameraMove::Move()
     auto& cameraTransform = GetComponent<Tabby::TransformComponent>();
     auto& camera = GetComponent<Tabby::CameraComponent>();
 
-    cameraTransform.position.x += velocity.x;
-    cameraTransform.position.y += velocity.y;
+    camera.Camera.target = { cameraTransform.Position.x, cameraTransform.Position.y, cameraTransform.Position.z - 1 };
+
+    cameraTransform.Position.x += velocity.x;
+    cameraTransform.Position.y += velocity.y;
     // playerRB.SetVelocity({ velocity.x, playerRB.GetVelocity().y });
 
     // std::cout << "X: " << cameraTransform.Position.x << "  Y: " << cameraTransform.Position.y << "\n";
 
     // std::cout << GetFPS() << "\n";
-    camera.camera.target = { cameraTransform.position.x, cameraTransform.position.y, cameraTransform.position.z - 1 };
 
     if (IsKeyDown(KEY_T)) {
-        cameraTransform.rotation.x += 2;
+        cameraTransform.Rotation.x += 2;
         // std::cout << cameraTransform.Rotation.x << "\n";
     }
     if (IsKeyDown(KEY_Y)) {
-        cameraTransform.rotation.y += 2;
+        cameraTransform.Rotation.y += 2;
         // std::cout << cameraTransform.Rotation.y << "\n";
     }
 }