+++ Tabby/src/Tabby.h
@@ -1,7 +1,5 @@
 #include <imgui/imgui.h>
 #include <raylib.h>
-// #include "rayli
-// #include "raylib/src/raylib.h"
 
 #include <Core/Application.h>
 #include <Graphics/Animation.h>