+++ Tabby/src/tbpch.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <algorithm>
+#include <array>
 #include <cassert>
 #include <cmath>
 #include <cstdint>
@@ -10,13 +11,11 @@
 #include <memory>
 #include <optional>
 #include <random>
+#include <sstream>
+#include <string>
 #include <tuple>
 #include <unordered_map>
 #include <utility>
-
-#include <array>
-#include <sstream>
-#include <string>
 #include <vector>
 
 #include "Core/Assert.h"